requirements.txt
requests
openai
python-dotenv
langgraph
langsmith
typing_extensions
groq
anthropic
google-generativeai
streamlit
jira
pytest
pytest-json-report
python-decouple
